requirements.txt
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
instaloader==4.5.5
Jinja2==2.11.2
psycopg2==2.8.6
requests==2.25.1
tqdm==4.60.0
urllib3==1.26.2
virtualenv==20.2.2
Werkzeug==1.0.1
wget==3.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
scipy==1.6.3
Pillow==8.2.0
numpy==1.19.2
pandas==1.1.3